fix: replace uint256 for uint96

+++ packages/contracts/contracts/FeeDistributor.sol
@@ -85,7 +85,7 @@ contract FeeDistributor is CheckContract, FeeDistributorStorage, IFeeDistributor
         // Send fee to the SOVFeeCollector address
         uint256 feeToSovCollector = toDistribute.mul(FEE_TO_SOV_COLLECTOR).div(LiquityMath.DECIMAL_PRECISION);
         zusdToken.approve(address(sovFeeCollector), feeToSovCollector);
-        sovFeeCollector.transferTokens(address(zusdToken), feeToSovCollector);
+        sovFeeCollector.transferTokens(address(zusdToken), uint96(feeToSovCollector));
 
         // Send fee to ZERO staking contract
         uint256 feeToZeroStaking = toDistribute.sub(feeToSovCollector);
@@ -101,7 +101,7 @@ contract FeeDistributor is CheckContract, FeeDistributorStorage, IFeeDistributor
         uint256 feeToSovCollector = toDistribute.mul(FEE_TO_SOV_COLLECTOR).div(LiquityMath.DECIMAL_PRECISION);
         wrbtc.deposit{value: feeToSovCollector}();
         wrbtc.approve(address(sovFeeCollector), feeToSovCollector);
-        sovFeeCollector.transferTokens(address(wrbtc), feeToSovCollector);
+        sovFeeCollector.transferTokens(address(wrbtc), uint96(feeToSovCollector));
 
         // Send the ETH fee to the ZERO staking contract
         uint256 feeToZeroStaking = toDistribute.sub(feeToSovCollector);
